+++ lib/src/column_sort.dart
@@ -1,0 +1,9 @@
+import 'package:easy_table/src/sort_type.dart';
+
+/// Describes the column sort.
+class ColumnSort {
+  ColumnSort({required this.columnIndex, required this.sortType});
+
+  final int columnIndex;
+  final EasyTableSortType sortType;
+}