+++ lib/src/internal/header_cell.dart
@@ -1,6 +1,5 @@
 import 'package:axis_layout/axis_layout.dart';
 import 'package:easy_table/easy_table.dart';
-import 'package:easy_table/src/sort_type.dart';
 import 'package:flutter/material.dart';
 import 'package:meta/meta.dart';
 
@@ -12,12 +11,14 @@ class EasyTableHeaderCell<ROW> extends StatefulWidget {
       {Key? key,
       required this.model,
       required this.column,
-      required this.resizable})
+      required this.resizable,
+      required this.multiSortEnabled})
       : super(key: key);
 
   final EasyTableModel<ROW> model;
   final EasyTableColumn<ROW> column;
   final bool resizable;
+  final bool multiSortEnabled;
 
   @override
   State<StatefulWidget> createState() => _EasyTableHeaderCellState();
@@ -47,15 +48,24 @@ class _EasyTableHeaderCellState extends State<EasyTableHeaderCell> {
         shrink: theme.expandableName ? 0 : 1,
         expand: theme.expandableName ? 1 : 0));
 
-    if (widget.model.sortedColumn == widget.column) {
+    final EasyTableSortType? sortType = widget.column.sortType;
+    if (sortType != null) {
       IconData? icon;
-      if (widget.model.sortType == EasyTableSortType.ascending) {
+      if (sortType == EasyTableSortType.ascending) {
         icon = theme.ascendingIcon;
-      } else if (widget.model.sortType == EasyTableSortType.descending) {
+      } else if (sortType == EasyTableSortType.descending) {
         icon = theme.descendingIcon;
       }
       children.add(
           Icon(icon, color: theme.sortIconColor, size: theme.sortIconSize));
+      if (widget.model.isMultiSorted) {
+        children.add(Align(
+            alignment: Alignment.center,
+            child: Text(widget.column.sortOrder.toString(),
+                style: TextStyle(
+                    color: theme.sortIconColor,
+                    fontSize: theme.sortOrderSize))));
+      }
     }
 
     Widget header = AxisLayout(
@@ -70,13 +80,17 @@ class _EasyTableHeaderCellState extends State<EasyTableHeaderCell> {
     if (sortable) {
       header = MouseRegion(
           cursor: enabled ? SystemMouseCursors.click : MouseCursor.defer,
-          child: GestureDetector(
-              behavior: HitTestBehavior.opaque,
-              child: header,
-              onTap: enabled
-                  ? () => _onHeaderPressed(
-                      model: widget.model, column: widget.column)
-                  : null));
+          child: Focus(
+              onKeyEvent: (node, event) {
+                return KeyEventResult.handled;
+              },
+              child: GestureDetector(
+                  behavior: HitTestBehavior.opaque,
+                  child: header,
+                  onTap: enabled
+                      ? () => _onHeaderPressed(
+                          model: widget.model, column: widget.column)
+                      : null)));
     }
 
     if (resizable) {
@@ -148,17 +162,20 @@ class _EasyTableHeaderCellState extends State<EasyTableHeaderCell> {
 
   void _onHeaderPressed(
       {required EasyTableModel model, required EasyTableColumn column}) {
-    if (model.sortedColumn == null) {
+    if (model.isSorted == false) {
       model.sortByColumn(column: column, sortType: EasyTableSortType.ascending);
+    } else if (widget.multiSortEnabled) {
+      widget.model.multiSortByColumn(widget.column);
     } else {
-      if (model.sortedColumn != column) {
+      final EasyTableSortType? sortType = widget.column.sortType;
+      if (sortType == null) {
         model.sortByColumn(
             column: column, sortType: EasyTableSortType.ascending);
-      } else if (model.sortType == EasyTableSortType.ascending) {
+      } else if (sortType == EasyTableSortType.ascending) {
         model.sortByColumn(
             column: column, sortType: EasyTableSortType.descending);
       } else {
-        model.removeColumnSort();
+        model.clearSort();
       }
     }
   }